web-summarizer-backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
requests==2.31.0
beautifulsoup4==4.12.3
google-generativeai==0.5.0
python-dotenv==1.0.1
python-docx==1.1.0
PyMuPDF==1.23.7